Completed phase 1 - Pull and Analyze Financial Statement (Vertical & Horizontal Analysis)

+++ app/Http/Controllers/SymbolController.php
@@ -12,9 +12,12 @@ use Illuminate\Http\Request;
 use App\Models\FinancialStatement;
 use App\Jobs\PullFinancialStatement;
 use Bkstar123\BksCMS\AdminPanel\Role;
+use App\Http\Components\RequestByUserThrottling;
 
 class SymbolController extends Controller
 {
+    use RequestByUserThrottling;
+
     /**
      * Display a list of financial statements
      *
@@ -28,10 +31,7 @@ class SymbolController extends Controller
             if (!auth()->user()->hasRole(Role::SUPERADMINS)) {
                 $financial_statements = $financial_statements->where('admin_id', auth()->user()->id);
             }
-            $financial_statements = $financial_statements->simplePaginate(config('bkstar123_bkscms_adminpanel.pageSize'))
-                                                         ->appends([
-                                                            'search' => $searchText
-                                                        ]);
+            $financial_statements = $financial_statements->simplePaginate(config('bkstar123_bkscms_adminpanel.pageSize'))->appends(['search' => $searchText]);
         } catch (Exception $e) {
             $financial_statements = [];
         }
@@ -46,22 +46,27 @@ class SymbolController extends Controller
      */
     public function pullFinancialStatement(Request $request)
     {
-        $request->validate([
-            'symbol' => 'required',
-            'year' => 'required|integer|between:1900,2100',
-            'quarter' => 'required|integer|between:0,4'
-        ]);
-        $data = $request->except('_token');
-        $data['admin_id'] = $request->user()->id;
-        try {
-            $financialStatement = FinancialStatement::create($data);
-            PullFinancialStatement::dispatch($request->except('_token'), $financialStatement->id, $request->user());
-            flashing('Your request is being processed')
-            ->flash();
-        } catch (Exception $e) {
-            flashing('Failed to proceed the requested action')
-            ->error()
-            ->flash();
+        if (!$this->isThrottled()) {
+            $this->setRequestThrottling();
+            $request->validate([
+                'symbol' => 'required',
+                'year' => 'required|integer|between:1900,2100',
+                'quarter' => 'required|integer|between:0,4'
+            ]);
+            $data = $request->except('_token');
+            $data['admin_id'] = $request->user()->id;
+            try {
+                $financialStatement = FinancialStatement::create($data);
+                PullFinancialStatement::dispatch($request->except('_token'), $financialStatement->id, $request->user());
+                flashing('Your request is being processed')
+                ->flash();
+            } catch (Exception $e) {
+                flashing('Failed to proceed the requested action')
+                ->error()
+                ->flash();
+            }
+        } else {
+            flashing('KSTOCK is busy processing your first request, please wait for 10 seconds before sending another one')->warning()->flash();
         }
         return back();
     }
@@ -74,15 +79,20 @@ class SymbolController extends Controller
      */
     public function destroyFinancialStatement(FinancialStatement $financial_statement)
     {
-        try {
-            $financial_statement->delete();
-            flashing("The selected financial statement has been successfully removed")
+        if (!$this->isThrottled()) {
+            $this->setRequestThrottling();
+            try {
+                $financial_statement->delete();
+                flashing("The selected financial statement has been successfully removed")
                 ->success()
                 ->flash();
-        } catch (Exception $e) {
-            flashing("The submitted action failed to be executed due to some unknown error")
+            } catch (Exception $e) {
+                flashing("The submitted action failed to be executed due to some unknown error")
                 ->error()
                 ->flash();
+            }
+        } else {
+            flashing('KSTOCK is busy processing your first request, please wait for 10 seconds before sending another one')->warning()->flash();
         }
         return back();
     }
@@ -94,16 +104,21 @@ class SymbolController extends Controller
      */
     public function massiveDestroyFinancialStatements()
     {
-        $Ids = explode(',', request()->input('Ids'));
-        try {
-            FinancialStatement::destroy($Ids);
-            flashing('All selected financial statements have been removed')
+        if (!$this->isThrottled()) {
+            $this->setRequestThrottling();
+            $Ids = explode(',', request()->input('Ids'));
+            try {
+                FinancialStatement::destroy($Ids);
+                flashing('All selected financial statements have been removed')
                 ->success()
                 ->flash();
-        } catch (Exception $e) {
-            flashing("The submitted action failed to be executed due to some unknown error")
+            } catch (Exception $e) {
+                flashing("The submitted action failed to be executed due to some unknown error")
                 ->error()
                 ->flash();
+            }
+        }  else {
+            flashing('KSTOCK is busy processing your first request, please wait for 10 seconds before sending another one')->warning()->flash();
         }
         return back();
     }
